F Operations\ add sweetalert in handleDelete func

+++ frontend/src/pages/Operations/index.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 import OperationTable from '../../components/OperationsTable';
 import { baseUrl } from '../../config';
 import './operations.css';
@@ -17,9 +18,26 @@ function Operations() {
     getOperations();
   }, []);
   const handleDelete = async (Id) => {
-    fetch(`${baseUrl}/operations/${Id}`, {
-      method: 'DELETE',
+    const result = await Swal.fire({
+      title: '¿Estas seguro?',
+      text: 'No se va a poder revertir esta acción',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminar',
     });
+    if (result.isConfirmed) {
+      await fetch(`${baseUrl}/operations/${Id}`, {
+        method: 'DELETE',
+      });
+      Swal.fire(
+        'Eliminado!',
+        'La operación ha sido eliminada.',
+        'success',
+      );
+    }
+
   };
   return (
     <div className='mt-5'>
